+++ indico/core/signals/event/registration.py
@@ -110,3 +110,15 @@ Executed after a registration form is cloned. The sender is the old ``Registrati
 object being cloned. The new ``RegistrationForm`` object is passed in the ``new_form``
 kwarg.
 ''')
+
+registration_form_section_can_access = _signals.signal('registration-form-section-can-access', '''
+Called when generating registration form sections data.  The `sender` is the
+``RegistrationForm`` object.  The section object is passed in the `section`
+kwarg and the registration object is passed in the `registration` kwarg.  The
+`management` kwarg is set to `True` if the registration form is rendered /
+processed from the event management area. If the access check fails, the section
+will be skipped.
+
+If multiple subscribers to the signal return contradictory results, ``False``
+wins and access is denied.
+''')